:lipstick: Tweaking the layout of the map
It's not there yet, I can't quite get it to where I want; but this is
getting closer. I want the map to fit in a specific area of screen and
scroll in x and y as and when needed. Can't for the life of me figure out ow
to do that though and I've just spent the last goodness-knows-how-long
trying to make that happen.